Move dataset classes into their own submodule
