+++ setup.py
@@ -176,6 +176,7 @@ setup(name = 'veusz',
                      ],
       packages = [ 'veusz',
                    'veusz.dataimport',
+                   'veusz.datasets',
                    'veusz.dialogs',
                    'veusz.document',
                    'veusz.helpers',
