+++ veusz/dialogs/datacreate.py
@@ -22,6 +22,7 @@ from ..compat import cstr
 from .. import qtall as qt4
 from .. import utils
 from .. import document
+from .. import datasets
 from .veuszdialog import VeuszDialog
 
 def _(text, disambiguation=None, context="DataCreateDialog"):
@@ -120,7 +121,7 @@ class DataCreateDialog(VeuszDialog):
         """Given a dataset name, allow it to be edited again
         (if it is editable)."""
 
-        if isinstance(ds, document.DatasetExpression): 
+        if isinstance(ds, datasets.DatasetExpression): 
             # change selected method
             if ds.parametric is None:
                 # standard expression
@@ -142,7 +143,7 @@ class DataCreateDialog(VeuszDialog):
                     text = ''
                 self.dsedits[part].setText(text)
 
-        elif isinstance(ds, document.DatasetRange):
+        elif isinstance(ds, datasets.DatasetRange):
             # change selected method
             self.valueradio.click()
             # make sure name is set
@@ -228,7 +229,7 @@ class DataCreateDialog(VeuszDialog):
             self.statuslabel.setText(status)
 
         except (document.CreateDatasetException,
-                document.DatasetException, _DSException) as e:
+                datasets.DatasetException, _DSException) as e:
 
             # all bad roads lead here - take exception string and tell user
             if dsexists:
@@ -279,8 +280,8 @@ class DataCreateDialog(VeuszDialog):
             vals[key] = (minval, maxval)
             
         linked = self.linkcheckbox.checkState() == qt4.Qt.Checked
-        return document.OperationDatasetCreateRange(name, numsteps, vals,
-                                                    linked=linked)
+        return document.OperationDatasetCreateRange(
+            name, numsteps, vals, linked=linked)
 
     def createParametric(self, name):
         """Use a parametric form to create the dataset.
@@ -299,9 +300,8 @@ class DataCreateDialog(VeuszDialog):
                 vals[key] = text
 
         linked = self.linkcheckbox.checkState() == qt4.Qt.Checked
-        return document.OperationDatasetCreateParameteric(name,
-                                                          t0, t1, numsteps,
-                                                          vals, linked=linked)
+        return document.OperationDatasetCreateParameteric(
+            name, t0, t1, numsteps, vals, linked=linked)
 
     def createFromExpression(self, name):
         """Create a dataset based on the expressions given."""
